Refactor enemy classes
