+++ src/scenes/GameScene.js
@@ -1,6 +1,6 @@
 import Phaser from "phaser";
-import Zombie from "./Zombie.js";
-import Skeleton from "../classes/Skeleton.js";
+import Zombie from "../classes/Enemies/Zombie.js";
+import Skeleton from "../classes/Enemies/Skeleton.js";
 import Player from "../classes/Player";
 import Bullet from "../classes/Bullet";
 import assets from "../../public/assets";
